Added command line functionality.
Removed all the hardcoded stuff about sport and the weeks, so that you
can use the following format to call it anywhere. Helps with setting up
the schedule through the crontab.

dmsw.rb SPORT WEEK

So for example:
  dmsw.rb fb 12
will look for the 12th week of the football season.

  dmsw.rb bb 4
will look for the 4th week of the bb season (well, month, really due to
the way ESPN handles their web page scores)

Removed a few files and functions that weren't used any more.